+++ customized-cms/cms-main/CHANGELOG.md
@@ -25,7 +25,10 @@
     - Responsive styling *look for 'June 2025 C&O styling ' comments*
       - removing fancy styling like box shadow, transient etc for smaller screens 
     - Hard coding special lists into list.ts files instead of dynamically compiling them Comments 'June 2025 C&O hard code lists'
+      - 'June 2025 C&O restructuring'
       - Creating 'latest-uploads' (*Done for '-timewise'*), 'best', 'changers' and other folders under 'special-lists'
+      - 'latest-uploads-timewise' changed to 'most-recently-uploaded' and move up in menu tree
+      - 'latest-uploads-themewise' put in cold storage - stopped making general sense and too much to maintain; too little ROI
       - Creating the individual lists like 'latest-uploaded-timewise.image.list.ts' (*Done*) under the respective folder
       - Initial population using 'console.log(JSON.stringify()) in an array from the compiled list', then copy paste into the .list.ts file
       - Right now loading is semi-automated using 'Write JSON to file' duplicate entry